Move controller into home folder Set routes in consequence Set specs in consequence Change link in consequence
