+++ app/controllers/administration/items_controller.rb
@@ -36,7 +36,7 @@ module Administration
       @item = Item.find(params[:id])
       CartItem.where(item_id: @item.id).each { |cart_item| CartItem.delete(cart_item) }
       CategoryItem.where(item_id: @item.id).each { |category_item| CategoryItem.delete(category_item) }
-      # OrderItem.where(item_id: @item.id).each {|order_item| OrderItem.delete(order_item) }
+      ItemOrder.where(item_id: @item.id).each { |order_item| ItemOrder.delete(order_item) }
       Item.delete(@item)
       redirect_to request.referer
     end
