--- app/controllers/category_items_controller.rb
@@ -1,4 +1,0 @@
-# frozen_string_literal: true
-
-class CategoryItemsController < ApplicationController
-end
