+++ app/controllers/home/category_items_controller.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+module Home
+  class CategoryItemsController < ApplicationController
+  end
+end
