+++ app/views/administration/items/edit.html.erb
@@ -7,12 +7,12 @@
 
    <div class="form-group">
          <%= label_tag 'Nom' %>
-         <%= f.text_field :name ,placeholder: @item.name ,class: "form-control "  %>
+         <%= f.text_field :name  ,class: "form-control "  %>
    </div>
 
    <div class="form-group">
          <%= label_tag 'Description' %>
-         <%= f.text_area :description, placeholder: @item.description ,class: "form-control "  %>
+         <%= f.text_area :description, class: "form-control "  %>
    </div>
 
    <div class="form-group">
