+++ app/views/home/carts/index.html.erb
@@ -6,7 +6,7 @@
 <%= item.reduced_price%> €
 <br>
 <% cart_item = CartItem.where(cart_id: @cart.id ,item_id: item.id)[0]%>
-<%= link_to "DELETE", cart_item_path(cart_item.id), :method => :delete,  data: { confirm: 'Are you sure you want to delete this article from your cart ?' } %>
+<%= link_to "DELETE", home_cart_item_path(cart_item.id), :method => :delete,  data: { confirm: 'Are you sure you want to delete this article from your cart ?' } %>
 <br><br>
 <% end %>
 
@@ -14,5 +14,5 @@
 
 <br>
 <% if !@cart.items.empty? %>
-<h3><%= link_to "Passer la commande", orders_path, method: "post" %></h3>
+<h3><%= link_to "Passer la commande", home_orders_path, method: "post" %></h3>
 <% end%>
