+++ app/views/home/items/index.html.erb
@@ -34,7 +34,7 @@
 											<%=cat.name %>
 											<%end%>
 									</td>
-									<td><%= link_to "Ajouter au panier", cart_item_path(item.id), method: "patch"%></td>
+									<td><%= link_to "Ajouter au panier", home_cart_item_path(item.id), method: "patch"%></td>
 
                 </tr>
                 <% end%>
