+++ config/routes.rb
@@ -6,16 +6,21 @@ Rails.application.routes.draw do
 
   get '/landing', to: 'landing#landing_page'
 
-  resources :carts, only: %i[index destroy]
-  resources :cart_items, only: %i[update destroy]
-  resources :orders, only: %i[create]
+  # resources :carts, only: %i[index destroy]
+  # resources :cart_items, only: %i[update destroy]
+  # resources :orders, only: %i[create]
 
   namespace 'home' do
     get '/', to: 'categories#index'
-    resources :items, only: %i[show]
+
+    resources :carts, only: %i[index destroy]
+    resources :cart_items, only: %i[update destroy]
+    resources :orders, only: %i[create]
+
     resources :categories do
       resources :items, only: %i[index destroy]
     end
+
     resources :profiles
     resources :items, only: %i[index show]
   end
