+++ spec/controllers/home/carts_controller_spec.rb
@@ -2,6 +2,6 @@
 
 require 'rails_helper'
 
-RSpec.describe OrdersController, type: :controller do
+RSpec.describe Home::CartsController, type: :controller do
   pending "add some examples to (or delete) #{__FILE__}"
 end
